+++ webots_ros2_driver/src/plugins/static/Ros2Camera.cpp
@@ -153,13 +153,8 @@ namespace webots_ros2_driver {
       vision_msgs::msg::ObjectHypothesisWithPose hypothesis;
       hypothesis.pose.pose = pose.pose;
       detection.results.push_back(hypothesis);
-#if FOXY
-      detection.bbox.center.x = objects[i].position_on_image[0];
-      detection.bbox.center.y = objects[i].position_on_image[1];
-#else
       detection.bbox.center.position.x = objects[i].position_on_image[0];
       detection.bbox.center.position.y = objects[i].position_on_image[1];
-#endif
       detection.bbox.size_x = objects[i].size_on_image[0];
       detection.bbox.size_y = objects[i].size_on_image[1];
       mRecognitionMessage.detections.push_back(detection);
@@ -169,13 +164,8 @@ namespace webots_ros2_driver {
       recognitionWebotsObject.id = objects[i].id;
       recognitionWebotsObject.model = std::string(objects[i].model);
       recognitionWebotsObject.pose = pose;
-#if FOXY
-      recognitionWebotsObject.bbox.center.x = objects[i].position_on_image[0];
-      recognitionWebotsObject.bbox.center.y = objects[i].position_on_image[1];
-#else
       recognitionWebotsObject.bbox.center.position.x = objects[i].position_on_image[0];
       recognitionWebotsObject.bbox.center.position.y = objects[i].position_on_image[1];
-#endif
       recognitionWebotsObject.bbox.size_x = objects[i].size_on_image[0];
       recognitionWebotsObject.bbox.size_y = objects[i].size_on_image[1];
       for (size_t j = 0; j < objects[i].number_of_colors; j++) {