+++ webots_ros2/CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for package webots_ros2
 ------------------
 * Added new TIAGo project to webots_ros2_tiago to run real robot configuration.
 * Added new WebotsController node in the driver interface to launch robot controller plugins.
+* Fixed unfound robot window library in Tesla example.
 
 2023.0.4 (2023-XX-XX)
 ------------------