+++ webots_ros2/CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog for package webots_ros2
 
 2023.0.4 (2023-05-23)
 ------------------
+* Drop support for Foxy.
 * Fixed vertical field of view in static RangeFinder plugin.
 * Added support for painted point clouds
 * Fixed ability to launch RViz without other tools in e-puck example.